move data into its own modulw
